requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
gspread==5.10.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
google-generativeai==0.3.0
Pillow==10.0.0
stripe==5.5.0
python-dateutil==2.8.2
pytz==2023.3
werkzeug==2.3.7